+++ sdmx-dl-provider-ri/src/main/java/sdmxdl/provider/ri/http/HttpFactorySupport.java
@@ -17,7 +17,7 @@ import java.util.List;
  * </p>
  *
  * @see HttpFactory
- * @see UrlConnectionHttpFactory
+ * @see DefaultHttpFactory
  */
 @lombok.Builder
 public final class HttpFactorySupport implements HttpFactory {